Implemented a caching system to only compile the assets we need when we call rake assets:pack. Requires the presence of an asset-manifest.json file that the rake script will build.

+++ lib/tasks/assets.rake
@@ -5,15 +5,15 @@ namespace :assets do
 
 	OUTPUT_DIR = "public"
 
-	task :pack => ["assets:make_app_build_js", "assets:clean_copy", "assets:compile_react", "assets:run_r_js", "assets:generate_polyfill", "assets:uglify", "assets:compile_css", "assets:clean_output_dir", "assets:fake_assets", "assets:precompile", "assets:restore_assets", "assets:write_manifest"] #
+	task :pack => ["assets:make_app_build_js", "assets:clean_copy", "assets:run_r_js", "assets:generate_polyfill", "assets:uglify", "assets:clean_output_dir", "assets:fake_assets", "assets:precompile", "assets:restore_assets", "assets:write_manifest"]
 	
-	task :distribute => ["assets:make_app_build_js", "assets:clean_copy", "assets:compile_react", "assets:run_r_js", "assets:generate_polyfill", "assets:uglify", "assets:compile_css", "assets:clean_output_dir"]
-
 	# Run r.js on the clean copy of our assets directory:
 	task :run_r_js do
 		puts "Running task assets:run_r_js"
 		system("node vendor/assets/r.js/dist/r.js -o app.build.js baseUrl=assets-clean_copy/ appDir='' mainConfigFile=assets-clean_copy/main.js")
-		FileUtils.cp("vendor/assets/requirejs/require.js", "public/assets/javascripts/")
+		if Dir.glob("public/assets/require*.js").length == 0
+			FileUtils.cp("vendor/assets/requirejs/require.js", "public/assets/javascripts/")
+		end
 		
 		FileUtils.rm_r "assets-clean_copy"
 		FileUtils.rm "app.build.js"
@@ -21,12 +21,13 @@ namespace :assets do
 
 	task :clean_output_dir do
 		puts "Running assets:clean_output_dir"
-		files = Dir.glob("#{OUTPUT_DIR}/assets/javascripts/routes/**/*.js") + Dir.glob("#{OUTPUT_DIR}/assets/**/*.css") + ["#{OUTPUT_DIR}/assets/javascripts/require.js", "#{OUTPUT_DIR}/assets/javascripts/main.js", "#{OUTPUT_DIR}/assets/javascripts/polyfill.js"] 
-		Dir.glob("#{OUTPUT_DIR}/assets/**/*.js").each do |file|
+		files = ["#{OUTPUT_DIR}/assets/javascripts/require.js", "#{OUTPUT_DIR}/assets/javascripts/polyfill.js"]
+		files += JSON.parse(File.read("build-manifest.json")).map{ |x| "#{OUTPUT_DIR}/assets/javascripts/#{x}.js"}
+		(Dir.glob("#{OUTPUT_DIR}/assets/javascripts/**/*.js")).each do |file|
 			next if files.include? file
 			File.delete(file)
 		end
-	
+		
 		Dir.glob("#{OUTPUT_DIR}/assets/**/*").select{ |d| File.directory? d}.sort{ |a,b| b.count('/') <=> a.count('/') }.each{ |d| Dir.rmdir(d) if (Dir.entries(d) - %w[.. .]).empty?}
 	end
 	
@@ -47,36 +48,77 @@ namespace :assets do
 	
 	task :make_app_build_js do
 		require 'json'
+		require 'digest/md5'
 		config = JSON.parse(Erubis::Eruby.new(File.read('app/assets/javascripts/main.js.erb')).result(binding()).gsub(/(\t|\n)/,'').split('var requirejs_configuration = ')[1].split(';')[0].gsub("'",'"'))
-		config['modules'] = [
-			{
-				'name' => 'main'
-			}
-		]
+		
+		built_modules = []
+		
+		manifest = (!File.exists? "asset-manifest.json") ? {} : JSON.parse(File.read("asset-manifest.json"))
+		
+		config["modules"] = []
+		
+		build_mod = false
+		if manifest.has_key? "main.js"
+			if Dir.glob("public/assets/main*.js").length == 0
+				build_mod = true
+			else
+				manifest["main.js"].each{ |file, hash| if Digest::MD5.hexdigest(File.read(file)) != hash then build_mod = true end }
+			end
+		else
+			build_mod = true
+		end
+		
+		if build_mod
+			config['modules'].push(
+				{
+					'name' => 'main'
+				}
+			)
+			built_modules.push("main")
+		end
+		
 		Dir.glob("app/assets/javascripts/routes/**/*.js").each do |route_file|
 			
 			route_file.gsub!("app/assets/javascripts/","").gsub!(/\.js$/,"")
+			build_mod = false
 			
-			mod_def = {
-				'name' => route_file,
-				'exclude' => ['main']
-			}
-
-			# Include the sidebar component in the build, if it exists:
-			if File.exists? "app/assets/javascripts/components/sidebar/#{route_file}.js.jsx" or File.exists? "app/assets/javascripts/components/sidebar/#{route_file}.js"
-				mod_def['include'] = [
-					"components/sidebar/#{route_file}"
-				]
+			if manifest.has_key? route_file+".js"
+				if Dir.glob("public/assets/#{route_file}*.js").length == 0
+					build_mod = true
+				else
+					manifest[route_file+".js"].each{ |file, hash| if Digest::MD5.hexdigest(File.read(file)) != hash then build_mod = true end }
+				end
+			else
+				build_mod = true
 			end
-			config['modules'].push(mod_def)
-
+			
+			if build_mod
+				mod_def = {
+					'name' => route_file,
+					'exclude' => ['main']
+				}
+				# Include the sidebar component in the build, if it exists:
+				if File.exists? "app/assets/javascripts/components/sidebar/#{route_file}.js.jsx" or File.exists? "app/assets/javascripts/components/sidebar/#{route_file}.js"
+					mod_def['include'] = [
+						"components/sidebar/#{route_file}"
+					]
+				end
+				config['modules'].push(mod_def)
+				
+				built_modules.push(route_file)
+			end
+			
 		end
+		
+		File.open("build-manifest.json", "w") do |fp|
+			fp.write(JSON.pretty_generate(built_modules))
+		end
+		
 		config['appDir'] = './assets-clean_copy'
 		config['baseUrl'] = 'assets'
 		config['dir'] = "./public/assets/javascripts"
 		config['skipDirOptimize'] = true
 		config['optimize'] = 'none'
-		
 		
 		File.open('app.build.js', 'w') do |json_fp|
 			json_fp.write(JSON.pretty_generate(config))
@@ -101,11 +143,15 @@ namespace :assets do
 
 	task :uglify do
 		require 'uglifier'
+		require 'json'
 	
 		puts "Running task assets:uglify"
 		
-		files = Dir.glob("#{OUTPUT_DIR}/assets/javascripts/routes/**/*.js") + ["#{OUTPUT_DIR}/assets/javascripts/require.js", "#{OUTPUT_DIR}/assets/javascripts/main.js", "#{OUTPUT_DIR}/assets/javascripts/polyfill.js"]
+		files = ["#{OUTPUT_DIR}/assets/javascripts/require.js", "#{OUTPUT_DIR}/assets/javascripts/polyfill.js"]
+		files += JSON.parse(File.read("build-manifest.json")).map{ |x| "#{OUTPUT_DIR}/assets/javascripts/#{x}.js"}
+		#files = Dir.glob("#{OUTPUT_DIR}/assets/javascripts/routes/**/*.js") + ["#{OUTPUT_DIR}/assets/javascripts/require.js", "#{OUTPUT_DIR}/assets/javascripts/main.js", "#{OUTPUT_DIR}/assets/javascripts/polyfill.js"]
 		files.each do |file|
+			next if not File.exists? file
 			puts "Uglifying #{file}"
 			compressed_source = Uglifier.compile(File.read(file))
 			File.open(file, 'w') do |f_pointer|
@@ -115,16 +161,18 @@ namespace :assets do
 	end
 
 	task :generate_polyfill do
-		puts "Running task assets:generate_polyfill"
-		polyfill = [
-			'/vendor/assets/respond/dest/respond.src.js',
-			'/vendor/assets/css3-mediaqueries-js/css3-mediaqueries.js',
-			'/vendor/assets/html5shiv/dist/html5shiv.js'
-		]
-		FileUtils.mkdir_p "public/assets/javascripts/"
-		File.open('public/assets/javascripts/polyfill.js', 'w') do |poly_fp|
-			polyfill.each do |p|
-				poly_fp.write(IO.read "./#{p}")
+		if Dir.glob("public/assets/polyfill-*.js").length == 0
+			puts "Running task assets:generate_polyfill"
+			polyfill = [
+				'/vendor/assets/respond/dest/respond.src.js',
+				'/vendor/assets/css3-mediaqueries-js/css3-mediaqueries.js',
+				'/vendor/assets/html5shiv/dist/html5shiv.js'
+			]
+			FileUtils.mkdir_p "public/assets/javascripts/"
+			File.open('public/assets/javascripts/polyfill.js', 'w') do |poly_fp|
+				polyfill.each do |p|
+					poly_fp.write(IO.read "./#{p}")
+				end
 			end
 		end
 	end
@@ -140,6 +188,10 @@ namespace :assets do
 	
 		# Read in the r.js build script:
 		requirejs_config = JSON.parse(IO.read('app.build.js').gsub(/^\(/,"").gsub(/\)$/,"").gsub(/\'/,"\""))
+		
+		if requirejs_config['modules'].length > 0
+			Rake::Task["assets:compile_react"].invoke
+		end
 	
 		dirs = Dir.glob('vendor/assets/*') + ['app/assets/javascripts']
 		requirejs_config['paths'].each do |k, file|
@@ -166,49 +218,52 @@ namespace :assets do
 		end
 	end
 
-	task :compile_css do
-		require 'compass'
-		require 'compass/sass_compiler'
-		puts "Running task assets:compile_css"
-		compiled_css =  Dir.glob('app/assets/stylesheets/**/[^_]*.scss').map{ |x| File.basename(x).gsub(/\.(?:scss|sass)/,'.css') }
-		FileUtils.mkdir_p "#{OUTPUT_DIR}/assets/stylesheets/"
-		compiled_css.each{ |file| File.unlink("#{OUTPUT_DIR}/assets/stylesheets/#{file}") if File.exists? file }
-		puts "Running Compass:"
-		Compass.add_configuration({
-			:http_path => "/",
-			:additional_import_paths => Dir.glob('vendor/assets/*'),
-			:css_dir => "#{OUTPUT_DIR}/assets/stylesheets",
-			:sass_dir => "app/assets/stylesheets",
-			:output_style => :compressed,
-			:line_comments => false
-		
-		}, "compass")
-		Compass.sass_compiler.compile!
-		Dir.glob("#{OUTPUT_DIR}/assets/stylesheets/*.css.css").each do |file|
-			FileUtils.mv(file, file.gsub(/\.css\.css$/,".css"))
-		end
-	end
-	
 	task :fake_assets do
-		FileUtils.mv("app/assets", "app/assets.proper")
+		require 'json'
+		FileUtils.mv("app/assets/javascripts", "app/javascripts.assets.proper")
 		FileUtils.mv("vendor/assets", "vendor/assets.proper")
-		FileUtils.mv("public/assets", "app/assets")
+		FileUtils.mkdir("app/assets/javascripts")
+		if(Dir.glob('public/assets/build*.txt').length > 0)
+			Dir.glob('public/assets/build*.txt').each do |f|
+				File.delete f
+			end
+		end
+		FileUtils.cp_r("public/assets/javascripts", "app/assets/")
+		FileUtils.rm_r("public/assets/javascripts")
+				
+		(!File.exists? "asset-manifest.json") ? {} : JSON.parse(File.read("asset-manifest.json")).each do |asset,requirements|
+			if asset =~ /\.js$/
+				asset_dir_in_question = "app/assets/javascripts"
+				asset_extension_in_question = "js"
+			else
+				next
+			end
+			
+			if not File.exists? "#{asset_dir_in_question}/#{asset}"
+				FileUtils.mkdir_p(File.dirname("#{asset_dir_in_question}/#{asset}"))
+				FileUtils.cp(Dir.glob("public/assets/#{asset.sub(/\.(js|css)$/,"")}*.#{asset_extension_in_question}").first, "#{asset_dir_in_question}/#{asset}")
+			end
+		end
+		
 	end
 	task :restore_assets do
-		FileUtils.rm_r("app/assets")
-		FileUtils.mv("app/assets.proper", "app/assets")
+		FileUtils.rm_r("app/assets/javascripts")
+		FileUtils.mv("app/javascripts.assets.proper", "app/assets/javascripts")
 		FileUtils.mv("vendor/assets.proper", "vendor/assets")
 	end
 	task :write_manifest do
 		require 'json'
 		require 'uglifier'
+		require 'digest/md5'
+		
 		manifest = {:files => {}, :assets => {}}
 
 		Dir.glob("public/assets/**/*.*").each do |file|
+			next if file =~ /manifest.json$/
 			mtime = File.mtime(file).to_s
 			size = File.size(file)
 			file.gsub!("public/assets/","")
-			basename = file.split(/-[a-f0-9]+\.[a-z]+$/)[0] + file.match(/\.[a-z]+$/)[0]
+			basename = file[0..file.rindex("-")] + file[file.rindex(".")..-1]
 			manifest[:files][file] = {
 				:logical_path => basename,
 				:digest => file.match(/-([a-f0-9]+)\.[a-z0-9]+$/)[1],
@@ -220,10 +275,36 @@ namespace :assets do
 		File.open("public/assets/manifest.json", "w") do |fp|
 			fp.write JSON.pretty_generate(manifest)
 		end
-		#main = IO.read("public/assets/#{manifest[:assets]['main.js']}")
-		File.open("public/assets/#{manifest[:assets]['main.js']}", "a") do |fp|
-			fp.write "#{Uglifier.compile("window.manifest = " + JSON.pretty_generate(manifest))};"
-			#fp.write main
+		
+		build_manifest = JSON.parse(File.read("build-manifest.json"))
+		
+		if(build_manifest.include? "main")
+			File.open(Dir.glob("public/assets/main*.js").first, "a") do |fp|
+				fp.write "#{Uglifier.compile("window.manifest = " + JSON.pretty_generate(manifest))};"
+			end
 		end
+		
+		require 'json'
+		require 'digest/md5'
+
+		asset_dirs = Dir.glob("vendor/assets/*") + ["app/assets/javascripts"]
+
+		tmp_manifest = {}
+		build = File.read(Dir.glob("public/assets/build*.txt").first).sub(/^\n/,"").split(/\n\n/).map{ |mod| (target, requirements) = mod.split("\n----------------\n"); tmp_manifest[target] = requirements.split("\n")}
+
+		asset_manifest = (!File.exists? "asset-manifest.json") ? {} : JSON.parse(File.read("asset-manifest.json"))
+
+		tmp_manifest.each { |file, mod| 
+			asset_manifest[file] = {} if not asset_manifest.has_key? file
+	
+			mod.each{|f|
+				actual_f = asset_dirs.map{ |d| if File.exists?("#{d}/#{f}") then "#{d}/#{f}" elsif File.exists?("#{d}/#{f}.erb") then "#{d}/#{f}.erb" elsif File.exists?("#{d}/#{f}.jsx") then "#{d}/#{f}.jsx" else nil end}.delete_if{|x| x.nil?}.first
+				asset_manifest[file][actual_f] = Digest::MD5.hexdigest(File.read(actual_f))
+			}
+		}
+		
+		File.open("asset-manifest.json", "w") { |fp| fp.write(JSON.pretty_generate(asset_manifest)) }
+		File.delete(Dir.glob("public/assets/build*.txt").first)
+		File.delete("build-manifest.json")
 	end
 end
